SubmitRequest: Fix access to uninitialized lateinit variable.
lateinit is read in the catch block, but is not yet initialized
on some exception paths. Since isInitialized() is not supported
on local lateinit variables, this PR convert it to a nullable
type so that we can check for initalization.

Since the BatchProofSigner ctor (obviously) creates a non-null
value, it should still be sufficiently clear that null can
_only_ arise if the ctor has not yet been called.